feat. like 컴포넌트 함수, 훅 분리
